backend/requirements.txt
flask
pymongo
mysql-connector-python
langchain
openai
python-dotenv
